break out sort proc into separate sources now that it has its own package
