+++ proc/sort/sort.go
@@ -1,25 +1,14 @@
 package sort
 
 import (
-	"bufio"
-	"container/heap"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"strconv"
 	"sync"
 
 	"github.com/brimsec/zq/ast"
 	"github.com/brimsec/zq/expr"
-	"github.com/brimsec/zq/pkg/fs"
 	"github.com/brimsec/zq/proc"
 	"github.com/brimsec/zq/zbuf"
-	"github.com/brimsec/zq/zio"
-	"github.com/brimsec/zq/zio/zngio"
 	"github.com/brimsec/zq/zng"
-	"github.com/brimsec/zq/zng/resolver"
 )
 
 // MemMaxBytes specifies the maximum amount of memory that each sort proc
@@ -235,211 +224,4 @@ func GuessSortKey(rec *zng.Record) string {
 		return fld
 	}
 	return "ts"
-}
-
-// runManager manages runs (files of sorted records).
-type RunManager struct {
-	runs       []*runFile
-	runIndices map[*runFile]int
-	compareFn  expr.CompareFn
-	tempDir    string
-	zctx       *resolver.Context
-}
-
-// NewRunManager creates a temporary directory.  Call Cleanup to remove it.
-func NewRunManager(compareFn expr.CompareFn) (*RunManager, error) {
-	tempDir, err := ioutil.TempDir("", "zq-sort-")
-	if err != nil {
-		return nil, err
-	}
-	return &RunManager{
-		runIndices: make(map[*runFile]int),
-		compareFn:  compareFn,
-		tempDir:    tempDir,
-		zctx:       resolver.NewContext(),
-	}, nil
-}
-
-func (r *RunManager) Cleanup() {
-	for _, run := range r.runs {
-		run.closeAndRemove()
-	}
-	os.RemoveAll(r.tempDir)
-}
-
-// CreateRun creates a new run containing the records in recs.
-func (r *RunManager) CreateRun(recs []*zng.Record) error {
-	expr.SortStable(recs, r.compareFn)
-	index := len(r.runIndices)
-	filename := filepath.Join(r.tempDir, strconv.Itoa(index))
-	runFile, err := newRunFile(filename, recs, r.zctx)
-	if err != nil {
-		return err
-	}
-	r.runIndices[runFile] = index
-	heap.Push(r, runFile)
-	return nil
-}
-
-// Peek returns the next record without advancing the reader.  The record stops
-// being valid at the next read call.
-func (r *RunManager) Peek() (*zng.Record, error) {
-	if r.Len() == 0 {
-		return nil, nil
-	}
-	return r.runs[0].nextRecord, nil
-}
-
-// Read returns the smallest record (per Less) from among the next records in
-// each run.  It implements the merge operation for an external merge sort.
-func (r *RunManager) Read() (*zng.Record, error) {
-	for {
-		if r.Len() == 0 {
-			return nil, nil
-		}
-		rec, eof, err := r.runs[0].read()
-		if err != nil {
-			return nil, err
-		}
-		if eof {
-			r.runs[0].closeAndRemove()
-			heap.Pop(r)
-		} else {
-			heap.Fix(r, 0)
-		}
-		if rec != nil {
-			return rec, nil
-		}
-	}
-}
-
-func (r *RunManager) Len() int { return len(r.runs) }
-
-func (r *RunManager) Less(i, j int) bool {
-	v := r.compareFn(r.runs[i].nextRecord, r.runs[j].nextRecord)
-	switch {
-	case v < 0:
-		return true
-	case v == 0:
-		// Maintain stability.
-		return r.runIndices[r.runs[i]] < r.runIndices[r.runs[j]]
-	default:
-		return false
-	}
-}
-
-func (r *RunManager) Swap(i, j int) { r.runs[i], r.runs[j] = r.runs[j], r.runs[i] }
-
-func (r *RunManager) Push(x interface{}) { r.runs = append(r.runs, x.(*runFile)) }
-
-func (r *RunManager) Pop() interface{} {
-	x := r.runs[len(r.runs)-1]
-	r.runs = r.runs[:len(r.runs)-1]
-	return x
-}
-
-// runFile represents a run as as a readable file containing a sorted sequence
-// of records.
-type runFile struct {
-	file       *os.File
-	nextRecord *zng.Record
-	zr         zbuf.Reader
-}
-
-// newRunFile writes sorted to filename and returns a runFile that reads the
-// file using zctx.
-func newRunFile(filename string, sorted []*zng.Record, zctx *resolver.Context) (*runFile, error) {
-	f, err := fs.Create(filename)
-	if err != nil {
-		return nil, err
-	}
-	r := &runFile{file: f}
-	if err := writeZng(f, sorted); err != nil {
-		r.closeAndRemove()
-		return nil, err
-	}
-	if _, err := f.Seek(0, 0); err != nil {
-		r.closeAndRemove()
-		return nil, err
-	}
-	zr := zngio.NewReader(bufio.NewReader(f), zctx)
-	rec, err := zr.Read()
-	if err != nil {
-		r.closeAndRemove()
-		return nil, err
-	}
-	return &runFile{
-		file:       f,
-		nextRecord: rec,
-		zr:         zr,
-	}, nil
-}
-
-// closeAndRemove closes and removes the underlying file.
-func (r *runFile) closeAndRemove() {
-	r.file.Close()
-	os.Remove(r.file.Name())
-}
-
-// read returns the next record along with a boolean that is true at EOF.
-func (r *runFile) read() (*zng.Record, bool, error) {
-	rec := r.nextRecord
-	if rec != nil {
-		rec = rec.Keep()
-	}
-	var err error
-	r.nextRecord, err = r.zr.Read()
-	eof := r.nextRecord == nil && err == nil
-	return rec, eof, err
-}
-
-// writeZng writes records to w as a zng stream.
-func writeZng(w io.Writer, records []*zng.Record) error {
-	bw := bufio.NewWriter(w)
-	zw := zngio.NewWriter(bw, zio.WriterFlags{})
-	for _, rec := range records {
-		if err := zw.Write(rec); err != nil {
-			return err
-		}
-	}
-	if err := zw.Flush(); err != nil {
-		return nil
-	}
-	return bw.Flush()
-}
-
-type unseenFieldTracker struct {
-	unseenFields map[ast.FieldExpr]expr.FieldExprResolver
-	seenTypes    map[*zng.TypeRecord]bool
-}
-
-func newUnseenFieldTracker(fields []ast.FieldExpr, resolvers []expr.FieldExprResolver) *unseenFieldTracker {
-	unseen := make(map[ast.FieldExpr]expr.FieldExprResolver)
-	for i, r := range resolvers {
-		unseen[fields[i]] = r
-	}
-	return &unseenFieldTracker{
-		unseenFields: unseen,
-		seenTypes:    make(map[*zng.TypeRecord]bool),
-	}
-}
-
-func (u *unseenFieldTracker) update(rec *zng.Record) {
-	if len(u.unseenFields) == 0 || u.seenTypes[rec.Type] {
-		return
-	}
-	u.seenTypes[rec.Type] = true
-	for field, fieldResolver := range u.unseenFields {
-		if !fieldResolver(rec).IsNil() {
-			delete(u.unseenFields, field)
-		}
-	}
-}
-
-func (u *unseenFieldTracker) unseen() []ast.FieldExpr {
-	var fields []ast.FieldExpr
-	for f := range u.unseenFields {
-		fields = append(fields, f)
-	}
-	return fields
 }
